Don't print errors if there are no nephtmp dirs in clt()

+++ bashrc
@@ -342,6 +342,7 @@ ct()
 clt()
 {
   for x in /tmp/nephtmp.*; do
+    [ -d "$x" ] || continue
     if [ -z "$(fuser "$x")" ]; then
       echo ":: Removing $x"
       rm -r $x
